[Impeller] disable vivante
